refactor: SelectableCard コンポーネントを CardSlot から CardModal 直下に移動
